requirements.txt
Flask==2.2.3
opencv-python==4.11.0.86
tensorflow==2.17.0
Pillow==9.2.0
numpy==1.24.2
Gunicorn==20.1.0